Move userinformation to main package
