+++ sdk/appcenter/src/main/java/com/microsoft/appcenter/utils/context/AbstractTokenContextListener.java
@@ -5,6 +5,8 @@
 
 package com.microsoft.appcenter.utils.context;
 
+import com.microsoft.appcenter.UserInformation;
+
 /**
  * Empty implementation to make callbacks optional.
  */
